Plan 133: usize/isize removed from Nova type system.
  - nova_int: int64_t -> intptr_t (signed address-sized, Go C-era intgo)
  - nova_uint: uintptr_t added (unsigned address-sized)
  - i64/u64 now separately mapped from int/uint
  - usize/isize -> E_TYPE_UNKNOWN with hint "use int"
  - 18 test files + std/raw_mem.nv (7 params) + vec_owned.nv migrated

Plan 134: ptr builtin removed; *() (pointer-to-unit) = void* in C codegen.
  - nova_ptr typedef removed from nova_rt.h
  - TypeRef::Pointer(Unit) -> "void*"; NullPtrLit -> ((void*)0)
  - ptr in type position -> error "use *()"
  - 4 codegen bugs fixed: void* tuple mono, sqlite_mini_ffi.h,
    TupleLit void* fallback, .0 field on void* newtype returning NULL
  - 20 files migrated (plan115x10, sync.nv, examples/ffi, plan118/plan131)

Conflict resolution: external unsafe fn syntax (main) + int param type (133/134)

+++ compiler-codegen/src/verify/pipeline.rs
@@ -2608,7 +2608,8 @@ fn type_ref_to_sort(ty: &crate::ast::TypeRef) -> SortRef {
                 "i16"          => SortRef::BitVec { width: 16, signed: true },
                 "u32"          => SortRef::BitVec { width: 32, signed: false },
                 "i32"          => SortRef::BitVec { width: 32, signed: true },
-                "u64" | "i32u" | "usize" => SortRef::BitVec { width: 64, signed: false },
+                // Plan 133: usize removed; uint = uintptr_t (64-bit on 64-bit targets).
+                "u64" | "i32u" | "uint" => SortRef::BitVec { width: 64, signed: false },
                 "bool" => SortRef::Bool,
                 "str"  => SortRef::Str,
                 "f32"  => SortRef::F32,
@@ -2888,7 +2889,8 @@ pub(super) fn type_to_sort(ty: &TypeRef) -> SortRef {
             "i16"          => SortRef::BitVec { width: 16, signed: true },
             "u32"          => SortRef::BitVec { width: 32, signed: false },
             "i32"          => SortRef::BitVec { width: 32, signed: true },
-            "u64" | "usize"            => SortRef::BitVec { width: 64, signed: false },
+            // Plan 133: usize removed; uint = uintptr_t (64-bit on 64-bit targets).
+            "u64" | "uint"             => SortRef::BitVec { width: 64, signed: false },
             "bool" => SortRef::Bool,
             "str"  => SortRef::Str,
             "f32"  => SortRef::F32,
